Shopping cart handling with react-query

+++ src/services/client/cart.service.js
@@ -1,10 +1,13 @@
 import { API_URL } from '@/config/config'
 import { axiosClient } from '@/lib/utils/axios'
+import { isDev } from '@/lib/utils/isDev'
 
 class CartService {
 	_CART = `${API_URL}/cart`
 
-	async getAllCartItems(userId) {
+	/* ----------------------------- Cart Items Ids ----------------------------- */
+	async getCartItemsIds(userId) {
+		if (!userId) return
 		const { data } = await axiosClient.get(this._CART, {
 			params: { userId }
 		})
@@ -12,7 +15,12 @@ class CartService {
 		return data
 	}
 
+	/* --------------------------- Add Product To Cart -------------------------- */
 	async addCartItem(userId, productId, category) {
+		if (!userId || !productId || !category) {
+			if (isDev()) console.error('Missing params')
+			return
+		}
 		await axiosClient.post(this._CART, {
 			userId,
 			productId,
@@ -20,10 +28,27 @@ class CartService {
 		})
 	}
 
+	/* ------------------------ Delete Product From Cart ------------------------ */
 	async deleteCartItem(userId, productId) {
+		if (!userId || !productId) {
+			if (isDev()) console.error('Missing params')
+			return
+		}
 		await axiosClient.delete(this._CART, {
 			params: { userId, productId }
 		})
+	}
+
+	/* ---------------------------- Cart Items Details --------------------------- */
+	async getCartItemsDetails(userId) {
+		if (!userId) {
+			if (isDev()) console.error('Missing userId')
+			return { data: [] }
+		}
+		const { data } = await axiosClient.get(`${this._CART}/detailed`, {
+			params: { userId }
+		})
+		return data
 	}
 }
 export const cartService = new CartService()
